Let increment and the new types work end to end
Increment does not declare a type of its own, so its node keeps the
default valueType and the inspector validated that instead of the
variable, which put an error on every increment node and stopped it
running.

Package export kept its own copy of the type rules, which still named
the replaced types, so a package holding an integer, float or keyboard
key variable could not be exported at all. It now uses the shared rules.

Also updates the two browser assertions that still expected the previous
type mismatch wording.

+++ tests/e2e/editor-shell.spec.ts
@@ -1604,7 +1604,7 @@ test("numeric fields autocomplete number variables and suspend literal stepping"
 	await expect(page.getByRole("heading", { name: "Verification" })).toBeVisible();
 	await expect(
 		page.locator("[data-verification-result] li").filter({
-			hasText: /Delay.*Variable "system_date" has type string; this field accepts numeric variables/i,
+			hasText: /Delay.*Variable "system_date" has type string; this field accepts float variables/i,
 		}),
 	).toBeVisible();
 });
